Add files via upload
The huge dataset used is available at this link : https://www.kaggle.com/datasets/kazanova/sentiment140 In this project, a model is trained with a huge dataset of 1.6 million tweets to predict new tweets whether it is a positive tweet/comment or a negative tweet. Logistic regression was used for better accuracy and it was above 81% for test data. We can predict new data that it is positive or negative just by using the trained_model.sav file.